Reorganize sparse-matrix types.
