+++ include/gendil/Algebra/SparseMatrixTypes/sparsematrixtypes.hpp
@@ -5,7 +5,7 @@
 // Shared sparse matrix action arithmetic policy helpers.
 #include "gendil/Algebra/SparseMatrixTypes/sparseapplyarithmetic.hpp"
 
-// A block sparse row (BSR) sparse matrix type.
+// A block sparse row (BSR) sparse matrix type and action.
 #include "gendil/Algebra/SparseMatrixTypes/bsrmatrix.hpp"
 
 // Algebraic scatter/gather wrapper for BSR sparse matrix application.
@@ -14,11 +14,11 @@
 // Raw duplicated unsorted COO triplet assembly buffer.
 #include "gendil/Algebra/SparseMatrixTypes/rawcootripletbuffer.hpp"
 
-// Canonical sorted and duplicate-reduced COO sparse matrix type.
+// Canonical sorted and duplicate-reduced COO sparse matrix type and action.
 #include "gendil/Algebra/SparseMatrixTypes/coomatrix.hpp"
 
-// Canonical compressed sparse row sparse matrix type.
+// Canonical compressed sparse row sparse matrix type and action.
 #include "gendil/Algebra/SparseMatrixTypes/csrmatrix.hpp"
 
-// Canonical compressed sparse column sparse matrix type.
+// Canonical compressed sparse column sparse matrix type and action.
 #include "gendil/Algebra/SparseMatrixTypes/cscmatrix.hpp"
